Build site in webroot directory (#7)
Previously, generateHtml.ps1 would output the content of index.html to
stdout. This only really worked for single page sites. This change is
the first of several that will address having a multi-page site.

A bit of wording I didn't like was cleaned up in the README as well.

Issues: #2