Implement deploy to Netlify
This will replace deploy to GitHub Pages, and the old GitHub Pages URL
will redirect to the Netlify site.

Netlify offers some advantages, the biggest being we can use a custom
domain that doesn't still need a `/pwsh-live-doc/` path after it. Also,
it'll be nice to deploy to a site that's fully owned by this project,
rather than a subdirectory in a larger site.

Links #2