Add page about null-coalescing operator double evaluation (#25)
Since this operator was only introduced in PowerShell 7, I also added a
`MinVersion` parameter to `OutputCode` so that we weren't just
generating error output for versions less than 7.

We also now only get the version for a particular exe once. Before we
would require the version for each exe every time `OutputCode` ran,
which slowed down the build.