Replace classes with PSCustomObject (#37)
The fact that you can't reload classes in a given session was making it
kinda difficult to debug. Tests and builds needed to run in their own
sessions, and as a result the $error variable wasn't available to figure
out exactly where an error occurred.

While PSCustomObject isn't as well integrated into the type system as
classes are, I still think it's the best choice going forward.