Add a page talking about printing to stderr in methods
This is a fairly large change. In addition to the new page, it enables
the example runner to collect stderr and improves the output style for
displaying it. It also changes GetExeVersion to return the version in
the directory path for PowerShell Core packages. This speeds up the dev
inner loop a lot.