Update stderr stream page with new GitHub Actions behavior
I'm really not sure what changed, but now the behavior of this page's
examples is almost the same in GitHub Actions as on my computer. The
only difference remaining is in method stderr where a function returning
a String will still print everything on GitHub Actions (without newlines
for some reason) but nothing on my computer.

Addresses #29