Only remove contents of webroot before build, not directory itself
Previously, if we had a static file server like Python's http.server
running in the webroot, we'd have to kill it before and build and
restart it once the build was over, since it was locking the webroot and
we'd be unable to delete that directory. Now that we're just deleting
the webroot contents, the file server doesn't have to be killed.