Get rid of category names in page URLs (#53)
There isn't really much risk of page URLs conflicting without the
category name. Plus, this method allows more flexibility to move pages
between categories or even get rid of the concept of categories
altogether.

This change also introduces a redirects file which is used to generate
placeholder HTML pages for pages that no longer exist. These pages use a
meta tag to redirect to the new URL, which is a server-independent way
of doing redirection. This is necessary, since links might exist to the
old pages with the category URLs (no matter how unlikely).