Use Netlify redirects rather than home-spun redirect files
The old method of redirects didn't work with Netlify pretty URLs. This
method only works on Netlify, but if we move to a new hosting provider
we can deal with that then.

Note that I think it's worth not relying on Netlify pretty URLs in
general and writing all links without .html extensions to start with.
However, the Python static file server I'm using for testing doesn't
allow this. It's something to think about for the future.